# myapi/swagger.yml
openapi: 3.0.0
info:
  title: "My API"
  description: "MP DEI API"
  version: "0.0.1"
servers:
  - url: "/api"
paths:
...
  /keyword-endpoint:
    post:
      operationId: "keyword_endpoint.create"
      tags:
        - "Keyword"
      summary: "Create a new keyword on the keyword table"
      requestBody:
        description: "Keyword to add"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                keyword:
                  type: string
              required:
                - id
                - keyword
      responses:
        "201":
          description: "Keyword created successfully"
        "400":
          description: "Invalid input"
    get:
      operationId: "keyword_endpoint.get_popular"
      tags:
        - "Keyword"
      summary: "Get top-10 popular keywords between given years"
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: integer
          description: "Start year (inclusive)"
        - name: endDate
          in: query
          required: true
          schema:
            type: integer
          description: "End year (inclusive)"
      responses:
        "200":
          description: "Successfully retrieved top-10 popular keywords"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    keywordName:
                      type: string
                    publicationsCount:
                      type: integer
...